feat: update gold entry handling with new date and price formats
- Increment version code and name in build.gradle.kts
- Refactor UpsertGoldScreen to include a date picker dialog and handle date selection
- Modify UpsertGoldUiState to store date as a Long timestamp
- Update UpsertGoldVM to manage date selection and submission logic
- Change GoldCreateGoldReq and GoldUpdateGoldReq to use Int for price and Double for grams
- Adjust Swagger and Go documentation to reflect updated request structures
- Update backend gold model to use uint64 for price and ensure date parsing is handled correctly

+++ backend/internal/domain/gold.go
@@ -13,7 +13,7 @@ type Gold struct {
 	WalletID  uint      `json:"wallet_id"`
 	Date      time.Time `json:"date"`
 	Grams     float64   `json:"grams"`
-	Price     float64   `json:"price"`
+	Price     uint64    `json:"price"`
 	Type      string    `json:"type"` // pure_gold or jewelry
 	Carat     float64   `json:"carat"`
 	Notes     string    `json:"notes"`
